Fix wait time not compiling lolrip

+++ src/main/java/com/thatredhead/redbot/command/impl/SubscriberCommands.java
@@ -167,9 +167,9 @@ public class SubscriberCommands extends CommandGroup {
     }
 
     public static class SubscriptionFeed {
-        public static final long MIN_WAIT =     5*60;
-        public static final long MAX_WAIT = 24*60*60;
-        public static final long START_WAIT =   5*60;
+        public static final int MIN_WAIT =     5*60;
+        public static final int MAX_WAIT = 24*60*60;
+        public static final int START_WAIT =   5*60;
 
         public static final JsonDeserializer<SubscriptionFeed> DESERIALIZER = (jsonElement, type, jsonDeserializationContext) -> {
